+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Add `stub`, `stub_const` and `stub_any_instance_of`.
 - Add support for `-I` in the CLI.
 
 ## [0.1.1] - 2024-08-20